refactor: extract realtime notification subscription from page.tsx
Move the 100-line realtime notification subscription handler into
useRealtimeNotifications hook. This is the most complex block in
page.tsx and the source of unread dot bugs.

page.tsx: 1219 → 1074 lines (-145)

The hook handles:
- squad_message/squad_mention → set hasUnread or auto-mark read
- friend_request → add to suggestions as incoming
- friend_accepted → move from suggestions to friends
- squad_invite, friend_check, check_tag → toast + reload
- All non-squad notifications → add to bell list
